backend/requirements.txt
# -*- coding: utf-8 -*-
# CZSC API后端依赖

# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# CZSC核心库
czsc>=0.9.0

# 数据处理
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.24.0

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pymysql>=1.1.0

# 数据验证
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 日志
loguru>=0.7.0

# HTTP客户端
httpx>=0.25.0
requests>=2.31.0

# 认证（直接使用 bcrypt，避免 passlib 与 bcrypt 4.1+ 不兼容）
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# 工具
python-dateutil>=2.8.0
pytz>=2023.3

# 开发工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
